Beginnings of mutability
Summary:
This begins the typechecking framework for mutability.

We typecheck mutability as a pass of the Typed AST, and keep track of the mutability of local variables using the local environment.
This establishes a mutability environment, which contains the list of locals and their mutability.
It also introduces passes on the typed AST to check and update the mutability environment on assignment.

List of things that aren't implemented:
- Mutable annotation at definition site (D6639448)
- Mutable annotation at callsite (D6639448)
- MutableReturn annotation at definition site
- MutableReturn annotation at callsite for methods(D6639448)
- freeze (D6638942)
- List assignments, inout
- All lambdas aren't mutable and cannot take mutable parameters or return mutables (probably by design)
- Checks for mutable, mutableReturn annotations to be classes only
- Using variables can probably be folded into this infrastructure as well

Reviewed By: vladima

Differential Revision: D6607222

fbshipit-source-id: 05c0e8255849cc77779b4d77171369cdb80ed3b2