Fast path in Slab::find()
Summary:
Most pointers point directly to the start of an object, even though
we support interior pointers. Check the common case before searching
backwards for the start bit.

When the check succeeds, we can avoid computing the object's size;
refactor the API to return HeapObject* instead of HdrBlock, since
we don't really need to know the size at this point. Downgrade a
few logging counters to count objects instead of bytes.

Reviewed By: swtaarrs

Differential Revision: D6655679

fbshipit-source-id: b2186c6a807706a4151fe0ef465e17f38defe58a