<<__Mutable>> on functions, method calls
Summary:
This diff implements <<__Mutable>> on function parameters, as well as on methods with the this type.

We do two checks:
On function calls, we check that arguments passed to a function with mutable parameters are mutable.
On method calls, we also check that if the method is mutable, the calling expression is also mutable.

Reviewed By: vladima

Differential Revision: D6639448

fbshipit-source-id: 7935ebe6638ddaf0394fc96db251ef7afe42faa7