Nuclide HHVM Debugger: Support SetVariableCommand - 22/N
Summary:
This diff adds support for setting variable values from the scopes window. Currently, only primitive values are supported, and changing the type of a variable is not supported. Users will be able to assign objects, properties, null, and change variable types and identities via running PHP commands in the console (as they do today) but not via the reflection code path.

There are two reasons for this - the first is in cases where the assignment requires constructing a new object, initializing something, parsing a array syntax, etc, we're going to need to interpret the RHS of the assignment as code and need the interpreter to do it. The second reason is when we add support for setting child properties on objects.  The SetVariableRequest only knows about the immediate parent of an object - so if I have `$x.foo.bar = 5`, we'll have a Variant containing the property "foo", and we don't know the FQN of the property path is $x.foo anymore - meaning we can't just turn this into a PHP string like `<php $x.foo.bar = 5; ?>` and hand it to the evaluator.

Reviewed By: rightparen

Differential Revision: D6548657

fbshipit-source-id: 1bcc807cfc0d71c70667dd2a5b56e6f5e2c10425