Nuclide HHVM Debugger: Add stdout and stderr hooks in server mode - 23/N
Summary: In server mode, we need to explicitly hook the stdout and stderr streams from PHP and redirect them to the debugger client's console. This is not needed in script mode, because in script mode the wrapper has the real stdout and stderr streams from HHVM as a child process.

Reviewed By: rightparen

Differential Revision: D6552220

fbshipit-source-id: 9a2da3cb1e2883a0c6be42975f5e54906813038e