Nuclide HHVM Debugger - Implement dummy evaluate - 21.2/ N
Summary:
This diff adds the ability for evaluate requests to happen on the dummy thread while the target is running. Since evaluates can invoke function calls, which can contain breakpoints, this diff adds the ability for the dummy thread to hit breakpoints.

The dummy request now has a request info, which can be obtained by the hook when the thread is the dummy. Breakpoints are now installed on the dummy thread as well, and stack traces can be obtained for the dummy.

The dummy request thread is typically not visible to the UX. When the dummy hits a bp, the thread is shown in the UX. When it finishes its eval, a thread exit message is sent to hide the dummy again.

Reviewed By: rightparen

Differential Revision: D6588867

fbshipit-source-id: 70db5d7c7d46ffef801c0c5cfb71feeb4a0c74cf