Fix refcount issues in gvn
Summary:
gvn was conservatively inserting IncRefs for instructions marked PRc,
ignoring the fact that not all replacements would be reached on all
paths.

This finds the places to insert IncRefs and DecRefs. In most cases,
refcount-opts will be able to kill everything it inserts (which it
couldn't before because the IncRefs generally weren't balanced).

Reviewed By: swtaarrs

Differential Revision: D6648654

fbshipit-source-id: cd7cd36018658e137b920693192491f101a7e9c6