Nuclide HHVM Debugger - Implement conditional breakpoints - 22/N
Summary:
This diff adds support for conditional breakpoints.

Upon setting a breakpoint, we parse and compile it into a unit. When the breakpoint is hit, we evaluate the condition PHP and if it returns true (either a boolean that is literally True, or an int that is > 0) we break. Otherwise we don't.

If the condition fails to parse, fails to compile or returns something that is not a bool or int, we tell the user and break anyway.

Conditions ARE allowed to call functions and trigger side effects (including assignments) but are NOT allowed to recursively hit breakpoints, for sanity reasons.

I am disabling hitConditions for now, will revisit later. Looking at the protocol I'm not 100% sure what the condition the user enters is supposed to look like.

Reviewed By: rightparen

Differential Revision: D6564100

fbshipit-source-id: efddae27bfcd08893c906c296003920316534b08