fix the double-nullable bug with 2-ary idx
Summary: This is the only remaining case in the code where we wrap an arbitrary type with `Toption` without first stripping away any potential `Toption` hiding in there. This should fix the double-nullable bug, barring some weird case with generic inference I am not seeing. The issue is that we say the return type is `?Tv`, but `Tv` itself may be nullable. We don't really know until we resolve `call`.

Reviewed By: jamesjwu

Differential Revision: D6597104

fbshipit-source-id: e224abf80e53112fc98f0d8f147441f1e026ab7b