Add vec_or_dict typehint
Summary:
Add support for a new typehint called vec_or_dict. As its name suggests, it
allows either vec or a dict. We resisted inventing such a type-hint, as it can
be misused to write dubiously typed code. However, we now need such a concept to
match varray_or_darray.

Reviewed By: mxw

Differential Revision: D6679051

fbshipit-source-id: 3165a0bb5000117a9e567c0f3f049caaa4ac7afb