Set JIT maturity to 100 once rate of TC increase is too small
Summary:
With semantic routing, some partitions generate a lot less code and so they were
never reaching JIT maturity of 100.  This diff uses warmupStatusString(), which
checks the rate at which the TC is increasing, to tweak JIT maturity
calculation, forcing it to 100 once the server is regarded as warmed up.  This
diff also adjusts Eval.JitWarmupMaxCodeGenRate to a more reasonable value.

Reviewed By: binliu19

Differential Revision: D6681807

fbshipit-source-id: b29cd23034bbb03acb5c0e1e43c7a03896cab551