Nuke EZC from orbit
Summary:
- Delete EZC
 - Delete ProxyArray
 - Delete bits of RefData for EZC
 - Convert dso_test to HNI

Now that the final release of HHVM with support for PHP 5 is out, this diff will remove the PHP 5 extension compatibility layer from master. These changes will not affect any current releases of HHVM, and are to allow us to begin making changes to the runtime that were structurally challenging due to the support for this compatibility layer.

Reviewed By: ricklavoie, swtaarrs, fredemmott, mxw, jano, alexeyt, edwinsmith

Differential Revision: D6680293

fbshipit-source-id: 3e55348293afb8f1637173084c8c72bb9ad16a50