Properly clean up MIterTable after stack overflow
Summary:
If we ended a request with live strong iterators (due to a fatal such
as stack overflow) MIterTable::clear() used remove_if with a lambda
that returned true. The result was that it decremented m_population
for every element; so unless every element was populated, m_population
went negative, and bad things started to happen in the next request.

This just adds a clear method to TlsPodBag and uses it from
MIterTable::clear()

Reviewed By: ricklavoie

Differential Revision: D6687980

fbshipit-source-id: 9ad7334c4408adc615472cb6f4b3afda76d96cd8