Mitigate performance implications of type-checking this
Summary:
This diff adds a tag to HHBBC types known to be instantiations of the current late static bound class.  This allows HHBBC to remove `this` return type checks in many cases, and to determine better types at some method call sites.  The diff also introduces a new set of operations on types that do not ignore the context tag information:
- `a.moreRefined(b)`  <- similar to `a.subtypeOf(b)`
- `equivalently_refined(a, b)`  <- similar to `a == b`

Reviewed By: markw65

Differential Revision: D5392940

fbshipit-source-id: 74b54db8daa2ae0daccbbbd2abda88fb7f279c10