Two-phase heap tracing: conservative then exact
Summary:
Background: Doing heap compaction correctly means identifying which
objects must be pinned (not moved) and which can be safely copied.
To safely copy an object we must be able to update all pointers to it,
and the best time to update pointers is when we find them, during the
trace phase. But, what if we conservatively scan a pointer to an object
that already was copied? Options are:
 1. two passes
 2. some way to "undo" a copy
 3. no copying until conservative scanning is done.

This diff implements approach 3.

Refactor heap tracing into two phases. In phase 1, do all the
conservative scanning of roots, and any heap objects we discover
which are also 100% conservatively scanned. Certain roots are
mixed; defer exactly-scanned pointers to phase 2.

In phase 2, process exactly-scanned pointers and heap objects,
and assert if any conservative ranges are discovered. Later,
this is where defragmentation (copying) would occur.

Kill the Pin bit in GCBits, since the phasing allows us to identify which
objects were pinned with just the mark bit.

Reviewed By: swtaarrs

Differential Revision: D5631530

fbshipit-source-id: 587a23afed1471343404e5a34cd64ca2d7fc24c1