Switch parse tree to inline records
Summary:
This changes the HackC parse tree to use inline records with the hope
that this will improve the performance of the compiler. There are still a few
places (esp. around coroutines) where we convert between inline records and
proper records. Getting rid of those requires another substantial round of
refactoring.

Differential Revision: D6692023

fbshipit-source-id: 673ade6755bd676798b3ccbfbd83a3f377892fd0