Fix a bug in store-elim
Summary:
In the presence of cycles, resolve_cycle will produce a store whose
source is a Mov. The source of the Mov was left as the input to one of
the partially available stores. The intent is that we will eventually
insert a phi, and replace the source of the Mov with that.

In some, very complex control flow cases, we can insert another phi
earlier, making this one redundant; in that case resolve_flat didn't
replace the input to the input to the Mov, which would typically leave
a use of the SSATmp that wasn't dominated by its def (and certainly
wasn't the correct value).

Instead this flags the Mov by setting its dst as its source, so that
even if we don't need a phi, we know to replace the source of the Mov
anyway.

Also adds lots of TRACING to this process to make it easier to debug
issues like this.

Reviewed By: mxw

Differential Revision: D6684539

fbshipit-source-id: dd27ff009bdccd582b25a030a52a90c0d2ec7f7a