Stop file I/O in CLI-server mode when client terminates
Summary: Adds a specific surprise-flag to handle client termination instead of relying on the existing one for timeouts. This further closes the gap between CLI-server mode and per-process HHVM instances for scripts by not having I/O happen after the "script" is killed. Particularly this stops the surprising behavior where we see stdio messages getting written to the terminal after script clients are terminated with ^C.

Reviewed By: paulbiss

Differential Revision: D6698971

fbshipit-source-id: 8cb7339206b6ed9837bc659226dd0965def5d87e