Fix for legacy compliance; `print` should be `echo`, but not as function
Summary:
For compliance with the legacy parser in `hh_server`, `print` should be parsed
as `echo`, but only when it's used as a unary prefix operator. When it's
applied as a function (i.e. the arguments are in parentheses), it should be
parsed as a function called `print`.

So
```
print "hello";
```
should be parsed as
```
echo "hello";
```
whereas
```
print("hello");
```
should not.

Reviewed By: vladima

Differential Revision: D6692776

fbshipit-source-id: 8299d65a60450956da808577f38ddfc67407c847