Kill the pessimized state in refcount opts
Summary:
I'm working on using the RCAnalysis in the partial-redundancy version of refcount opts. It actually plugged in quite nicely, and I was able to modify rcgraph_opts to abort if it did anything, run pre first, and have all the tests, and perflab pass (ie pre did a strict superset of what rcgraph_opts did). But perflab showed a .5-1% regression; and the reason was that pre-refcount opts using its simplistic analysis caught a lot of cases that the more sophisticated RCAnalysis missed.

The problem is that RCAnalysis will only track refcounts from direct incs, and memory support. And eg, if a lower_bound is zero, an IncRef will result in a lower_bound of 1, even though we know that the refcount must be at least 2 immediately after the IncRef. In addition, once we pessimize in a block, we force all lower_bounds to stick to zero until exit from the region.

This is the first of two diffs to try to rectify the situation. This one removes the pessimized flag, and the Halt nodes. As far as I can tell, this was just a jit-time performance consideration, on the grounds that there's really nothing we can optimize after such a node, and I think thats based on the assumption that downstream of an UnknownEffects, we can't ever have a balanced memory location. I see two issues there - first, after the next php call, everything is guaranteed to be balanced again; and second we can end up pessimizing just because we tried to lower a lower bound that was already zero (and in that case there's no reason there can't be balanced loads immediately).

I think we can safely drop the pessimized flag, and simply reset all lower_bounds to zero, and all memory support to empty in pessimize situations, and then let things run as normal. This isn't expected to be much of a win on its own, but clears the way to tracking "unsupported" refcounts, which I'll do in a follow up.

Reviewed By: mxw

Differential Revision: D6590413

fbshipit-source-id: 5b338f96d5ba3c1bf010cdb8c24cf39e10c2d388