Track unsupported refs in refcount-opts
Summary:
Following on from D6590413, this adds unsupported_refs to each must-alias set. These are used to say that we know there are references, but we don't know where they came from. This can happen eg if the lower_bound was zero, and we do an IncRef. We know the refcount is 2, but we don't have any support bits, so any instruction that might affect refcounts has to kill all the unsupported_refs.

This allows us to

  - remove IncRef/DecRef pairs with no non-irrelevant intervening instructions, even after pessimizing.
  - defer reducing the lower_bound when we see a PureStore, but don't know where its storing to.
  - always produce a non-zero lower_bound from DecRefNZ

This diff attempts to ensure that the results RCAnalysis are at least as good as the analysis done in remove_trivial_inc_decs. A follow up will eliminate all the pre-emptive decrefs, and limit the cases where we kill the unsupported_refs.

Reviewed By: ottoni

Differential Revision: D6590741

fbshipit-source-id: 0b4515b505b8a3280056a46a0a922f6e0a089a0e