Better memory support merging.
Summary:
Its always ok to drop support bits and replace them with
unsupported_refs, so do that when merging blocks.

I noticed that both the original algorithm (prior to unsupported_refs)
and my update to use unsupported_refs were not idempotent; merging the
same src into dst a second time could result in further changes. Thats
clearly not a good property to have, so I came up with this instead.

Reviewed By: edwinsmith

Differential Revision: D6671570

fbshipit-source-id: c8b6b5e9d99885ee36900fa847ff66b99148df83