Don't emit inout wrappers for abstract methods
Summary: This was causing issues during the conversion as abstract methods are not generally declared async, and so subclasses declaring async implementations of these methods were flagged as being incompatible with their base classes. I think we need a more general solution for wrappers when a subclass decides to make an overridden method async but this will suffice for the time being.

Reviewed By: ricklavoie

Differential Revision: D6707318

fbshipit-source-id: 4374983bf04279fa9c20e61e6176f2496a5f1e5b