Parity with legacy parser for dropping bodies in decl mode
Summary:
Previously, we ascribed the dropping of (function / method) bodies to
quickmode. `Parser_hack` has separate behaviour for files in `Mdecl` mode,
though. Going through the `Parser_hack` and identifying all the sites of this
leads to this attempt to synchronise behaviours. Note that we explicitly
exclude dropping based on `Mdecl` when we're in `hhvm_compat_mode`, because
codegen should always get all code.

Reviewed By: gregory-nisbet

Differential Revision: D6657872

fbshipit-source-id: 12a5b99be43fb5fd6e7e6319fb3a2e5573d7a1e6