Store tcopt in Tast.EnvAnnotation.t
Summary:
Currently, consumers of a Typed AST need to construct a Typing_env to make use of the TAST's EnvAnnotation. This means they must provide TypecheckerOptions, but there is no guarantee that the provided options are the same options used during typechecking.

To simplify the consumption of Typed ASTs, and to ensure that there is no mismatch between the options used during construction of the Typed AST and the later consumption of it, this change adds a tcopt field to the TAST's EnvAnnotation.

Reviewed By: andrewjkennedy

Differential Revision: D6595246

fbshipit-source-id: 5dccd608d449f51b18319ab38e9d25e4ab2d7549