Perform constant folding on Vptr operands
Summary:
In the vasm simplifier, perform constant folding on Vptr operands. For example,
if the base register is a constant, we can fold it into the displacement (as
long as the resultant value fits within 32-bits). We can do likewise for a
constant index. This is especially helpful for indexing into static arrays,
where the address of the array is a constant, and the index might be as well.

Reviewed By: markw65

Differential Revision: D6716329

fbshipit-source-id: 98cac936682db66e1f6ddc4c7aed3f184e0cbd3c