Handle unsupported_refs better in may_decref
Summary:
This is how I originally wanted to implement may_decref in the
presence of unsupported_refs; but I got a lot of crashes, and
convinced myself it was wrong (and the crashes went away when I
switched to the current version). But while updating the refcount opts
documentation, I realized that it ought to be correct; if we're
removing a "supported" reference, then that reference was exclusively
allocated to this must-alias set, so there's no need to decref
anything else.

I've made lots of other cleanup and fixes since, and this seems pretty
solid now in all my testing, so I'm assuming the crashes were down to
some other bug, since fixed.

Reviewed By: ricklavoie

Differential Revision: D6711815

fbshipit-source-id: 4c0e66f630bcd9abf85c447fea8e98facd535943