Re-order vasm operands in simplifier to allow for immediate form usage
Summary:
When emitting vasm instructions which have variations that can accept
immediates, its desirable to put a register corresponding to a constant as the
first operand. This is because the constant folder will automatically turn the
vasm instruction into its immediate taking form. Rather than open coding this
optimization every place we might emit the instruction, make the vasm simplifier
recognize this pattern. For certain vasm instructions which have an immediate
form (right now just the cmp* family), if the first operand is not a constant
register, but the second is, flip the operands (if safe to do so). Then the
constant folder can transform the instruction into the immediate form.

Reviewed By: markw65

Differential Revision: D6715913

fbshipit-source-id: e7578a08c122a034d387684d6a34505227be82eb