Nuclide HHVM Debugger - Put Server Constants and User Constants in a sub-scope - 26/N
Summary: Rather than having both types of constants as top-level containiers (on level with Locals and Superglobals) this diff moves them under a container named Constants. This gives us a HUGE perf win in Nuclide, because it asks for all the top level variables for each thread on pause - now the system constants (there are almost 3,000 of them) are requested lazily and only if the user expands that section.

Reviewed By: rightparen

Differential Revision: D6618022

fbshipit-source-id: 5917d0e025f4d51768379d66321a3df491b8fb25