Nuclide HHVM Debugger - Fix multiline stepping verbosity - 29/N
Summary:
Currently (and in XDebug as well) stepping over a multiline statement such as a function call that has arguments that are on multiple lines hits the end source line of the outermost statement multiple times. This is super annoying for the user.

Example:
1 foo_multiple(
2   'multiline statement',
3   bar(),
4 );

Hitting "next" would result in this hitting lines: 4, 2, 4, 3, 4 - after hitting 4 initially, you'd have to hit "next" 5 times to step over this.
The expected behavior would be to stop once at each argument only: 4, 2, 3 - requiring you to press "next" three times.

How annoying this is depends on how many arguments, but basically for each argument we have an extra step to the outer source location.

This diff fixes this to hit each actual line only once.

Reviewed By: rightparen

Differential Revision: D6630916

fbshipit-source-id: aa250ef1d550790c2ea02d41406ddd6c47ee376e