Nuclide HHVM Debugger - Include capabilities in launch/attach response message - 35/N
Summary:
This updates to include the debugger capabilities in the response to LaunchRequest and AttachRequest. While not actually documented (currently) in the VS Code debugger protocol, in practice, VS Code checks to see if these messages include capabilities, and re-loads the debugger capabilities in the editor based on the response body.

This is necessary to work around an issue in the protocol - the capabilities as spec'ed, are returned by InitializeResponse - which happens before launch/attach. In practice, there are features we may not know if they are supported or not until we've launched the backend and can ask it for its version, etc. Providing these again in the launch/attach response allows this dynamic enabling or disabling of features after HHVM is launched.

Reviewed By: rightparen

Differential Revision: D6670094

fbshipit-source-id: 3194d1d47ac28246a1fd9f6157d930760593441b