Nuclide HHVM Debugger - Allow attaching to already-running script - 36/N
Summary:
This change adds support to HHVM to attach to an already-running PHP script that was started with the `--mode vsdebug` and `--vsDebugPort <n>` arguments specified.

A new `--vsDebugNoWait` flag is added so that a script can startup and begin to run, but be attachable later. If this flag is absent, the default behavior is the "run in terminal" behavior, which blocks script startup until the debugger attaches.

Nuclide will show any running scripts with --vsDebugPort parameters, and allow attaching to them to begin debugging.

Reviewed By: rightparen

Differential Revision: D6680436

fbshipit-source-id: 30f26647d6134db4a0c24b37152790f91399e955