Nuclide HHVM Debugger - Ensure event messages are sent after command responses - 37/N
Summary:
This is a fix to our behavior around sending events (especially things like Breakpoint Events) during client commands. The comments in the VS protocol document specifically indicate ordering for a couple events, whereas others are left ambiguous.

When processing a SetBreakpointsRequest, for example, it is expected that the SetBreakpointsResponse goes out first, then the Breakpoint Event indicating bps were added, resolved, etc.

Thread Events are always going to go out immeidately - since new threads can come, go and stop at any time unrelated to the client command.

Finally, events are not deferred when performing an evaluation request, because hitting an inner breakpoint must be sent to the client immediately or we'll hang.

Reviewed By: rightparen

Differential Revision: D6689530

fbshipit-source-id: fc608bb951c3803db37f5dc8c566c65820c073d9