ban variable-variables in all modes from both parsers
Summary:
Variable-variables are "banned" in non-decl mode in the old parser, but the FFP doesn't care about them and parses them without any errors, afaict. I'd like to move this check to `Naming`, before we drop `Ast.Dollar` on the floor and turn it into `N.Any` (meaning we can't do this in `nastCheck`), that way this construct is banned no matter where we get the AST from.

I'd also like to ban this in the .hhconfig in the next release since it already was sort-of banned through parse error [1008] and being confident that it doesn't exist in the codebase will let us delete all the relevant HHVM code to support this feature as we fork from PHP.

Reviewed By: andrewjkennedy

Differential Revision: D6705828

fbshipit-source-id: 7181a6e46a67e64c522bf312d4e9866ed5dc9cd8