fix and simplify typechecking of ?-> (nullsafe get)
Summary:
Fixing bug of `?[unresolved]?->something` resolving to `[unresolved]`
instead of `?[unresolved]`.

This involves a simplification of `obj_get_`. Examining the
diff that introduced typing of ?-> (D1524942), we can see that the preexisting
continuation `k` is reused for wrapping the result in a `Toption` when the get is
nullsafe. This is really
not necessary and makes the code more complex than it needs to be. In D1524942, we
can see that the author actually attempted to simplify that part already, but introduced
a bug while doing so, therefore received a comment, and therefore undid the
simplification.

It also mislead someone
to add the comment at the top of obj_get_ in diff D2327282, when actually the use
of continuation dates back from before the introduction of ?->. I'm actually
not sure what the continuation is for but it has to do with that part of the
code: https://fburl.com/kosx78uv

Reviewed By: andrewjkennedy

Differential Revision: D6722918

fbshipit-source-id: 12fb61616579ce6f58e16491cd1d300a33e25de1