BigMalloc must store requested bytes for correct type_scan size.
Summary:
To correctly type_scan a req::malloc'd block, the size we pass
to the generated scanner function must be an integer multiple
of sizeof(T) for the object type being scanned.  For example, if
sizeof(T)==24, calling a generated scanner with size==32 triggers
an assert.

If a req::malloc'd block was rounded up to an even larger size,
then it's important to avoid calling the scanner with the rounded
up size; we can expect that the trailing bytes (possibly > sizeof(T))
are uninitialized.

It turns out we did this correctly for SmallMalloc, but not BigMalloc.
Refactor BigMalloc and BigObj to store the requested size in
MallocNode::nbytes, instead of the rounded up size. Take care to
still use the rounded up size for memory stats calculations.

At the same time, change the HeapImpl api slightly to only return void*
instead of MemBlock, since the callers really didn't need MemBlock.

Reviewed By: swtaarrs

Differential Revision: D6694467

fbshipit-source-id: fc567894e34f8d32ff97412cce569b9218843c5f