Nuclide HHVM Debugger - Suppress evaluation errors for hovers and data tips
Summary:
Nuclide doesn't send the evaluation context hint yet, but the backend should look for it and honor it if specified.

If provided, evaluation errors won't be printed for evals that are the result of watch expressions and data tips, so we don't get console spew when hovering over random tokens.

Reviewed By: rightparen

Differential Revision: D6724591

fbshipit-source-id: e8b26abf0011d561866ee441d461e0cfaa3d1f42