Rename variable names to avoid collisions
Summary:
When generating a closure class, it is important not to redeclare several names, such as `$nextLabel`, `$resume`, `$coroutineContinuation_generated`, `$clone`, etc. However, we are currently doing that when we generate object properties from used variables in coroutine functions. For instance, the following function produces an error:

```
coroutine function variable_names(): int {
  $stateMachineFunction = 1;
  $clone = suspend SuspendCoroutine::suspend(
    $cont ==> {
      $cont->resume(1);
    }
  );
  return $nextLabel
    + $stateMachineFunction;
}
```

```
Fatal error: Cannot redeclare Global_variable_names_GeneratedClosure::$stateMachineFunction in /data/users/thomasjiang/fbsource/fbcode/hphp/hack/test/facebook/coroutine/tests/test.php on line 1
```

To that end, instead of generating closure class properties directly from the variable name, instead, prepend all generated object properties with `"saved_"` so that they are guaranteed not to collide with any important fields, such as `$clone`.

Note that this not fix the issue that autogenerated closure parameters may be inadvertently trampled by the user. For instance, `$coroutineResult`, `$closure`, `$coroutineData`, `$exception_generated` are the four closure parameters that may be trampled if the user declares a variable of the same name in the coroutine. This fix will come in another diff.

Changes:
- Factor out `lstrip "$"` into coroutine_syntax.ml
- Change the `used_parameters` field in `Coroutine_context.t` to `function_parameters` as `used_parameters` was unused. This allows us to no longer pass around the function parameters as a separate argument into the closure generator since we already pass in the generated context.
- Added function to prepend "saved_" to variables
- Added test

Reviewed By: michaeltingley

Differential Revision: D6690492

fbshipit-source-id: 6bd7cdaf75977d432df1c85698aa74add5e2649e