ServerInferType: Make Tanon behavior more explicit
Summary: Previously, there was a typo here--we invoked super#on_expr when we should have invoked self#on_expr. Because super#on_expr returns None in most situations where an anonymous function is used, this seemed to have the right behavior, but it wasn't obvious why this was the case. This change explicitly checks for Tanon and Tunresolved types and adds a comment explaining why we return None for them.

Reviewed By: pittsw

Differential Revision: D6735489

fbshipit-source-id: 05876966becc519a1fa49f0e23b3c1e45c0fc2c8