Handle unresolved receivers when typing instance methods in TAST
Summary:
Previously, if a receiver was an unresolved union of multiple classes, the type of an Obj_get representing an instance method was the function type of that method on the class whose type appeared last in the Tunresolved list of the receiver's type.

After this change, we type the instance method as the union of the function types of each method that might be invoked.

Reviewed By: dlreeves

Differential Revision: D6735487

fbshipit-source-id: c881e867b664b4fd554f9e2ab97b4f4f52734b4b