Nuclide HHVM Debugger - Fix flaky test shutdown
Summary: When the HHVM process finishes running a script, it can exit without flushing the remaining VS debug protocol messages into the TCP socket. The real client doesn't care so much when this happens, since the debug wrapper terminates the debug session if it sees the TCP socket close. The tests should be resiliant to this as well.

Reviewed By: vladima

Differential Revision: D6733753

fbshipit-source-id: cbf87d46c37a4a1404f8ddf87d8087e180bffadc