Don't skip debugger tests
Summary:
There was a rule in hphp/test/DEFS to skip debugger or ext_vsdebug
tests when "jit" was in the mode. But mode is always either "jit" or
"interp,jit", so the tests were always skipped.

Lets see what happens if we don't skip them (they seem to run ok
locally).

Reviewed By: swtaarrs

Differential Revision: D6741022

fbshipit-source-id: a3e25d71e0000c940b1946b8760ba9fcd76f2b73