Replace hh_format with hackfmt for typechecker tests
Summary:
Replace `hh_format` with `hackfmt` for the typechecking tests and fix up any differences due to the different formatting behaviour.

`hh_format` is the legacy formatter, while `hackfmt` is the replacement. There's a question about whether the typechecker tests need a formatting test, however I've decided to leave them in as:  1) we already have them, 2) they provide additional coverage for `hackfmt` and that helps uncover bugs in `hackfmt`.

Reviewed By: CatherineGasnier

Differential Revision: D6702163

fbshipit-source-id: 13922c95831598566b657940d79b3313517a79bb