Clear line table cache on Unit destruction
Summary:
As-is, if a Unit is freed and another Unit is allocated at the same
memory address, if the cache line for that address has not been cleared
naturally by another Unit that hashes the same, we can end up looking up
a stale line table for an unrelated Unit.

Atomically clear the corresponding cache line in ~Unit() to avoid this.

Reviewed By: markw65

Differential Revision: D6749772

fbshipit-source-id: 76c44283a539272558d72bc2b8eabf83e849b246