Change the parameter field from set to list to retain order
Summary:
When collecting coroutine function parameters, they were being collected in a set, then extracted from the set. This meant that the function parameters were being ordered alphabetically when passed to the closure, but the function parameters are not meant to be ordered alphabetically.

For instance, the function declared as:

```
coroutine function parameter_ordering(
  int $b,
  vec<int> $z,
  MyObj $a,
  string $y,
): void {
```

But would have had the effect of the function parameters being passed in alphabetically.

Changes:
- Changed the set to a list so that we retain parameter ordering.
- Added a runtime test to ensure behavior is correct.

Reviewed By: michaeltingley

Differential Revision: D6764070

fbshipit-source-id: c4986624d5ef21ba55e25d8289e8cdd24bc7bcfa