Untemplateize refcount opts
Summary:
Most of the rc-analysis computation was templatized on an NAdder
parameter - but we only ever use two; one is a no-op, and the other is
PreAdder.

Just rewrite PreAdder so it has a no-op mode, and untemplatize
everything. This involved shuffling code around a bit so that
definitions precede uses.

Reviewed By: ricklavoie, swtaarrs

Differential Revision: D6771475

fbshipit-source-id: d8f5ebe069bd3ab5a58864472f57319e898b2e6b