Consider multiple smashables per target when relocating.
Summary:
When finding the smashable from its target, we need to consider all possible smashables instead of just stopping at the first match. The rationale is that the first possible smashable that we find might not be an actual smashable. Consider a case where a smashable `jmp` is coincidentally preceded by a conditional branch. This can make it look like a smashable `jcc` which is the first possible smashable found. Without considered all possible smashables, we'd miss this `jmp`.
Closes https://github.com/facebook/hhvm/pull/8057

Differential Revision: D6564085

Pulled By: mxw

fbshipit-source-id: 424f613c7e2b745c476dc1d3cb74f90129c02a77