Type alias cleanups.
Summary:
resolveTypeAlias shouldn't need to check for AnnotType::Self and AnnotType::Parent; type aliases can't have those (unlike type constants) because they can only be introduced at the top level and not in a class. The emitter already rejects attempts to do this.
Type aliases are not experimental syntax for some time now, so we can drop a bunch of .hphp_opts files.

Reviewed By: swtaarrs

Differential Revision: D6748794

fbshipit-source-id: 0b3dbe5ce0c88041f12f50014f6ec1cf6af059af