Strip strict typechecking
Summary:
When lowering coroutines, the classes generated often have untyped fields/parameters, which will fail to pass the typechecking test if the raised code/file is in strict mode.

In this diff, the code that we generate will strip the strict mode from the file, producing a partially typed lowered code file.

Reviewed By: michaeltingley

Differential Revision: D6756402

fbshipit-source-id: 5aa70962bf1bde18b531e22f920f4fda84eddf9f