Remove use of load script in integration tests
Summary:
The load script is no longer used anywhere except in these integration tests (where
we write a shell script that acts as a fake load script).
Before we delete it wholesale from the code, we need to remove this last usage.

This ports the integration tests to use the --with-mini-state flag instead.

Reviewed By: dabek

Differential Revision: D6754349

fbshipit-source-id: 3c467d4db93e3e2085e5058e03eedfbbe0849877