Don't produce errors when parameters in lambdas have no type hints
Summary:
Given a test script:

```
<?hh //strict

function test():void {
  $lambda = ($a, $b, $c, $d) ==> {
    return 1;
  };
}
```

the expected behaviour is to produce no errors due to the absence of type hints on the parameters, however the actual behaviour is to produce errors similar to:

```
(4,14)-(4,15) A type annotation is required in strict mode.
(4,18)-(4,19) A type annotation is required in strict mode.
(4,22)-(4,23) A type annotation is required in strict mode.
(4,26)-(4,27) A type annotation is required in strict mode.
```

Thie diff fixes that behaviour as well as adds a test for it.

Differential Revision: D6774327

fbshipit-source-id: 8cd90e7758dbbfce0e0edcab25ccdcacef22a2a3