-INF in adata
Summary:
Attributes in hhas files should be able to contain d:-INF which we now parse. Note
that although -INF should really be a constant, PHP treats it as 0 - INF so we only get it as
a literal in the output if we enable constant folding, which we don't currently do for Hack.

Also updated all the semdiff test expect files since size calculations were affected by
an earlier diff. And changed the test script to use consistent options when comparing with
and creating an expect file.

Reviewed By: andrewjkennedy

Differential Revision: D6760471

fbshipit-source-id: 229945ee439525f98ca4de73359078b8e5e57736